feat: add major number of types

+++ types/chat.go
@@ -1,7 +1,7 @@
 package types
 
 type Chat struct {
-	ID        int    `json:"id"`                   // Unique identifier for this chat. This number may have more than 32 significant bits and some programming languages may have difficulty/silent defects in interpreting it. But it has at most 52 significant bits, so a signed 64-bit integer or double-precision float type are safe for storing this identifier.
+	ID        int64  `json:"id"`                   // Unique identifier for this chat. This number may have more than 32 significant bits and some programming languages may have difficulty/silent defects in interpreting it. But it has at most 52 significant bits, so a signed 64-bit integer or double-precision float type are safe for storing this identifier.
 	Type      string `json:"type"`                 // Type of the chat, can be either “private”, “group”, “supergroup” or “channel”
 	Title     string `json:"title,omitempty"`      // Optional. Title, for supergroups, channels and group chats
 	Username  string `json:"username,omitempty"`   // Optional. Username, for private chats, supergroups and channels if available
